// estore/src/app/home/types/cart.type.ts
import { Product } from './products.type';

export interface CartItem {
  product: Product;
  quantity: number;
  amount: number;
}

export interface Cart {
  products: CartItem[];
  totalAmount: number;
  totalProducts: number;
}

export interface DeliveryAddress{
   userName: string;
    address: string;
    state: string;
    city: string;
    pin: string;
}